refactor: Main로직을 페이지별로 분리
